Starting to move mutation into its own file
